pyproject.toml
[project]
name = "dmp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "pedrogasg", email = "pedrogas_g@hotmail.com" }
]
requires-python = ">=3.14"
dependencies = [
    "camelot-py[cv]>=1.0.9",
    "docling>=2.74.0",
    "ghostscript>=0.8.1",
    "img2table[paddle]>=1.0.9",
    "polars<1.0",
    "pytesseract>=0.3.13",
    "thefuzz>=0.22.1",
]

[project.scripts]
parse = "dmp.cli:main"
json2mei = "dmp.exporters.json2mei:main"
extract-tables = "dmp.legacy.extract_tables:main"
straighten = "dmp.image.straighten:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.2",
    "ty>=0.0.18",
]
